requests in api are displayed in a separate module
