+++ index.html
@@ -43,6 +43,7 @@
     <script type="module" src="./render.js"></script>
     <script type="module" src="./reply.js"></script>
     <script type="module" src="./likeButton.js"></script>
+    <script type="module" src="./api.js"></script>
 
   </body>
 </html>
